refactor(rest): extract prepareImportRequest + CSV/xlsx parsers to import-prepare.ts (#2766 V2a)
Verbatim move out of rest-server.ts (module-private → exported) so the
identity import endpoint in plugin-auth can accept payloads byte-identical
to the generic /data/:object/import routes without re-implementing the
parsers. rest-server.ts imports the extracted module; the two call sites
and behavior are unchanged (import-integration tests as the regression
gate). Also exports the import building blocks (runImport, coerceRow,
buildFieldMetaMap, types) from the package index for cross-package reuse.
